+++ bin/actions/0220022-49.sh
@@ -1,9 +1,24 @@
 #!/bin/bash
 
-#: ◣ + N - Bring Nemo to front, launch if not yet running
+#: ◣ + N
+#: Chromium: Move active tab into new window
+#: Global: Bring Nemo to front, launch if not yet running
 
 focusApplication=\
 $(cat /proc/"$(xdotool getwindowpid "$(xdotool getwindowfocus)")"/comm)
+
+if [[ "$focusApplication" =~ "chromium-browse" ]] \
+|| [[ "$focusApplication" =~ "chromium" ]] \
+|| [[ "$focusApplication" =~ "firefox" ]]; then
+  xdotool sleep 0.1
+  xdotool key Control_L+l; xdotool sleep 0.1  # focus URL
+  xdotool key Control_L+c; xdotool sleep 0.3  # copy URL
+  xdotool key Control_L+w; xdotool sleep 0.3  # close tab
+  xdotool key Control_L+n; xdotool sleep 0.5  # open new window
+  xdotool key Control_L+v; xdotool sleep 0.1  # paste URL
+  xdotool key Return  # load URL
+  exit 0
+fi
 
 path_self="$( cd "$(dirname "$0")" >/dev/null 2>&1 || exit ; pwd -P )"
 "$path_self"/utils/remove_unwanted_output.sh "$focusApplication"